Refactor k-grid & k-column
