+++ panel/src/config/components.js
@@ -205,11 +205,11 @@ import Box from "@/ui/components/Box.vue";
 import Card from "@/components/Layout/Card.vue";
 import Cards from "@/components/Layout/Cards.vue";
 import Collection from "@/components/Layout/Collection.vue";
-import Column from "@/components/Layout/Column.vue";
+import Column from "@/ui/components/Column.vue";
 import Dropzone from "@/ui/components/Dropzone.vue";
 import Empty from "@/ui/components/Empty.vue";
 import FilePreview from "@/components/Layout/FilePreview.vue";
-import Grid from "@/components/Layout/Grid.vue";
+import Grid from "@/ui/components/Grid.vue";
 import Header from "@/components/Layout/Header.vue";
 import List from "@/components/Layout/List.vue";
 import ListItem from "@/components/Layout/ListItem.vue";
