+++ panel/src/ui/components/Grid.vue
@@ -1,11 +1,19 @@
 <template>
-  <div :data-gutter="gutter" class="k-grid">
+  <div
+    :data-gutter="gutter"
+    class="k-grid"
+  >
     <slot />
   </div>
 </template>
 
 <script>
 export default {
+  /**
+   * The grid does not have any gutter by default, but you can pass
+   * the following gutter sizes to control spacing between columns:
+   * `small`|`medium`|`large`|`huge`
+   */
   props: {
     gutter: String,
   }
@@ -14,25 +22,29 @@ export default {
 
 <style lang="scss">
 .k-grid {
+  --gap: 0;
   --columns: 12;
+  --max: 1fr;
+  --min: 12rem;
+  --col-gap: var(--gap);
+  --row-gap: var(--gap);
+
   display: grid;
-  grid-column-gap: 0;
-  grid-row-gap: 0;
+  grid-column-gap: var(--col-gap);
+  grid-row-gap: var(--row-gap);
   grid-template-columns: 1fr;
 }
 
 @media screen and (min-width: $breakpoint-sm) {
 
   .k-grid[data-gutter="small"] {
-    grid-column-gap: 1rem;
-    grid-row-gap: 1rem;
+    --gap: 1rem;
   }
 
   .k-grid[data-gutter="medium"],
   .k-grid[data-gutter="large"],
   .k-grid[data-gutter="huge"] {
-    grid-column-gap: 1.5rem;
-    grid-row-gap: 1.5rem;
+    --gap: 1.5rem;
   }
 
 }
@@ -42,19 +54,19 @@ export default {
     grid-template-columns: repeat(var(--columns), 1fr);
   }
   .k-grid[data-gutter="large"] {
-    grid-column-gap: 3rem;
+    --gap: 3rem;
   }
   .k-grid[data-gutter="huge"] {
-    grid-column-gap: 4.5rem;
+    --gap: 4.5rem;
   }
 }
 
 @media screen and (min-width: $breakpoint-lg) {
   .k-grid[data-gutter="large"] {
-    grid-column-gap: 4.5rem;
+    --gap: 4.5rem;
   }
   .k-grid[data-gutter="huge"] {
-    grid-column-gap: 6rem;
+    --gap: 6rem;
   }
 
 }
